+++ spec/integrations/setup/re_initialization_of_cached_resources_spec.rb
@@ -13,6 +13,7 @@ post_monitor = ::Karafka::Instrumentation::Monitor.new
 post_producer = Object.new
 
 Karafka::App.setup do |config|
+  config.kafka = { 'bootstrap.servers': 'host:9092' }
   config.logger = post_logger
   config.monitor = post_monitor
   config.producer = post_producer